More modifications to app structure, modify engine to support private members in older javascript environments, begin documentation of engine layer.
