+++ public/index.html
@@ -28,7 +28,6 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <script src="%PUBLIC_URL%/engine.js" type="text/javascript" />
     <div id="root"></div>
     <!--
       This HTML file is a template.
